% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BondHornKenny}
\alias{BondHornKenny}
\title{Triadic round robin data reported in Bond, Horn, and Kenny (1997)}
\format{
A data frame with 120 measurements from one round-robin group. A single number is a
judge's perception of an actor's liking of a partner. j.id is the id of the judge, p.id and 
t.id are the ids of the perceiver (actor) and target (partner), respectively.
}
\usage{
BondHornKenny
}
\description{
Data from Bond, Horn, and Kenny (1997) Table 3
}
\references{
Bond, C. F., Jr., Horn, E. M., & Kenny, D. A. (1997). A model for triadic relations. 
\emph{Psychological Methods}, \emph{2}(1), 79–94. 
\doi{10.1037/1082-989X.2.1.79}
}
\keyword{datasets}
