% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrm-methods.R
\name{coef.tsrm}
\alias{coef.tsrm}
\title{Extract Model Coefficients from a glmmX Object}
\usage{
\method{coef}{tsrm}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class `"glmmX"`.}

\item{...}{Currently not used.}
}
\value{
A named numeric vector of all model estimates.
}
\description{
Returns the estimated coefficients of a fitted TSRM or SRM.
}
\seealso{
[glmm()], [vcov.tsrm()], [summary.tsrm()]
}
