% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CurryEmerson}
\alias{CurryEmerson}
\title{Triadic round robin data from Curry and Emerson (1970)}
\format{
A data frame with 2,688 measurements from eight round-robin groups. A single number is a
judge's perception of an actor's liking of a partner.
}
\usage{
CurryEmerson
}
\description{
Data from a study by Curry and Emerson (1970) used in Kenny et al. (1996)
}
\references{
Kenny, D. A., Bond, C. F., Mohr, C. D., & Horn, E. M. (1996). Do we now how much people like
another? \emph{Journal of Personality and Social Psychology}, \emph{71}, 928–936. 
\doi{10.1037/0022-3514.71.5.928}
}
\keyword{datasets}
