% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HallmarkKenny}
\alias{HallmarkKenny}
\title{Round robin data from Hallmark and Kenny}
\format{
A data frame with 802 measurements from 30 round-robin groups on the following 7 round-robin
variables (taken on unnumbered 7-point rating scales with higher numbers indicating a
higher value of the trait): \cr

\code{calm}: rating of dimension calm-anxious \cr
\code{sociable}: rating of dimension sociable-withdrawn \cr
\code{liking}: rating of dimension like-do not like \cr
\code{careful}: rating of dimension careful-careless \cr
\code{relaxed}: rating of dimension relaxed-tense \cr
\code{talkative}: rating of dimension talkative-quiet \cr
\code{responsible}: rating of dimension responsible-undependable \cr

The data frame also contains the gender (\code{actor.sex}; \code{1} = F,
\code{2} = M) of the participants and their age in years (\code{actor.age}).
Note that the data was assessed in two conditions: odd round robin group numbers indicate
groups in which participants rated all traits for a person at a time whereas even numbers
refer to groups in which participants rated all the people for each trait.
}
\source{
\url{http://davidakenny.net/srm/srmdata.htm}
}
\usage{
HallmarkKenny
}
\description{
Data from Kenny et al. (1994)
}
\references{
Kenny, D. A., Albright, L., Malloy, T. E., & Kashy, D. A. (1994). Consensus in
interpersonal perception:  Acquaintance and the big five.
\emph{Psychological Bulletin, 116}(2), 245-258.
\doi{10.1037/0033-2909.116.2.245}
}
\keyword{datasets}
