% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsrm.R
\name{htsrm}
\alias{htsrm}
\title{Fit a bivariate Triadic Social Relations Model where the second (!) variable is a 
standard round-robin variable}
\usage{
htsrm(
  formula = NULL,
  g_var = NULL,
  p_var = NULL,
  data = NULL,
  control = make_args_list(),
  debug = FALSE
)
}
\arguments{
\item{formula}{A formula for the fixed effects}

\item{g_var}{A vector with an identifier for round-robin groups}

\item{p_var}{A vector with identifiers for actor and partner}

\item{data}{A data frame}

\item{control}{A list of control arguments}

\item{debug}{A control args for debugging, for internal use only}
}
\description{
Fit a bivariate Triadic Social Relations Model where the second (!) variable is a 
standard round-robin variable
}
