% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kenzer}
\alias{Kenzer}
\title{Zero acquaintance round robin data from David Kenny}
\format{
A data frame with 124 measurements from 7 round-robin groups on the following 5 round robin
variables (taken on unnumbered 7-point rating scales with higher numbers indicating a
higher value of the trait): \cr
\code{sociable}: rating of dimension sociable \cr
\code{irritable}: rating of dimension good-natured \cr
\code{responsible}: rating of dimension responsible \cr
\code{anxious}: rating of dimension calm \cr
\code{intellectual}: rating of dimension intellectual \cr
The data frame also contains the gender (\code{actor.sex}; \code{1} = F,
\code{2} = M) of the participants and their self-ratings on the five assessed traits
(\code{actor.sociable} and so on).
}
\source{
\url{http://davidakenny.net/srm/srmdata.htm}
}
\usage{
Kenzer
}
\description{
Data from a Albright et al. (1988) Study 2
}
\references{
Albright, L., Kenny, D. A., & Malloy, T. E. (1988). Consensus in 
personality judgments at zero acquaintance. \emph{Journal of Personality 
and Social Psychology}, \emph{55}(3), 387–395. 
\doi{10.1037/0022-3514.55.3.387}
}
\keyword{datasets}
