% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LashleyBond}
\alias{LashleyBond}
\title{Round robin data reported in Lashley and Bond (1997)}
\format{
A data frame with 30 measurements from one round-robin group. A single number is the
number of times an actor initiated aggression toward a partner.
}
\usage{
LashleyBond
}
\description{
Data from Lashley and Bond (1997) Table 1
}
\references{
Lashley, B. R., & Bond, C. F., Jr. (1997). Significance testing for round robin data. 
\emph{Psychological Methods}, \emph{2}(3), 278–291. 
\doi{10.1037/1082-989X.2.3.278}
}
\keyword{datasets}
