% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leipzig}
\alias{Leipzig}
\title{Zero acquaintance round robin data from Richard Rau}
\format{
A data frame with 2,162 measurements from one large round-robin group on the following 4 
round robin variables (taken on unnumbered 6-point rating scales with higher numbers indicating a
higher values): \cr
\code{liking}: rating of dimension liking \cr
\code{dominant}: rating of dimension dominant \cr
\code{affectionate}: rating of dimension affectionate \cr
\code{knowing}: rating of whether person i knows j \cr
}
\usage{
Leipzig
}
\description{
Data from a round robin study conducted by Richard Rau at the University of Leipzig in 2017.
}
\keyword{datasets}
