% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrm-methods.R
\name{summary.tsrm}
\alias{summary.tsrm}
\title{Summary method for tsrm objects}
\usage{
\method{summary}{tsrm}(object, digits = 3L, ...)
}
\arguments{
\item{object}{A fitted model object of class "tsrm"}

\item{digits}{Number of digits to display}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Summary method for tsrm objects
}
\seealso{
[glmm()], [coef.tsrm()], [vcov.tsrm()]
}
