% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrm.R
\name{tsrm}
\alias{tsrm}
\title{Fit a Triadic Social Relations Model to Generalized Round-Robin Data}
\usage{
tsrm(
  formula = NULL,
  p_var = NULL,
  g_var = NULL,
  data = NULL,
  control = make_args_list(),
  debug = FALSE
)
}
\arguments{
\item{formula}{A formula for the fixed effects}

\item{p_var}{A vector with identifiers for actor, partner, and judge}

\item{g_var}{A vector with an identifier for round-robin groups}

\item{data}{A data frame}

\item{control}{A list of control arguments}

\item{debug}{A control args for debugging, for internal use only}
}
\description{
Fit a Triadic Social Relations Model to Generalized Round-Robin Data
}
