% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrm-methods.R
\name{vcov.tsrm}
\alias{vcov.tsrm}
\title{Variance-Covariance Matrix for a tsrm Object}
\usage{
\method{vcov}{tsrm}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class `"tsrm"`.}

\item{...}{Currently not used.}
}
\value{
A symmetric numeric matrix. Row and column names match 
  [coef.tsrm()].
}
\description{
Returns the variance-covariance matrix of the estimated 
coefficients of a fitted TSRM or SRM.
}
\seealso{
[glmm()], [coef.tsrm()], [summary.tsrm()]
}
