% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Warner}
\alias{Warner}
\title{Round robin data reported in Warner et al. (1979)}
\format{
A data frame with 56 measurements from one round-robin group. Each variable represents the
proportion of time speaking by person i to j on one of three days.
}
\usage{
Warner
}
\description{
Data from a Warner et al. (1979) Table 7
}
\references{
Warner, R. M., Kenny, D. A., & Stoto, M. (1979). A new round robin 
analysis of variance for social interaction data. \emph{Journal of Personality 
and Social Psychology}, \emph{37}(10), 1742–1757.
\doi{10.1037/0022-3514.37.10.1742}
}
\keyword{datasets}
