% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kenzer}
\alias{kenzer}
\title{Example dataset for SRM fitting}
\format{
A data frame with 124 rows and 14 variables:
\describe{
  \item{rrgroup.id}{Id of the round-robin group}
  \item{actor.id}{Id of the actor (i.e., perceiver)}
  \item{target.id}{Id of the partner (i.e., target)}
  \item{sociable}{Sociable round-robin rating}
  \item{irritable}{Irritable round-robin rating}
  \item{responsible}{Responsible round-robin rating}
  \item{anxious}{Anxious round-robin rating}
  \item{intellectual}{Intellectual round-robin rating}
  \item{actor.sex}{Sex of actors}
  \item{actor.sociable}{Self-rating of sociable of actors}
  \item{actor.irritable}{Self-rating of irritable of actors}
  \item{actor.responsible}{Self-rating of responsible of actors}
  \item{actor.anxious}{Self-rating of anxious of actors}
  \item{actor.intellectual}{Self-rating of intellectual of actors}
}
}
\usage{
kenzer
}
\description{
Data from a round-robin study reported in Albright et al. (1988, Study 2).
}
\keyword{datasets}
