% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leipzig}
\alias{leipzig}
\title{Example dataset for SRM fitting}
\format{
A data frame with 2,162 rows and 6 variables:
\describe{
  \item{perceiver.id}{Id of the perceiver (actor)}
  \item{target.id}{Id of the target (partner)}
  \item{liking}{Liking round-robin rating}
  \item{dominant}{Dominant round-robin rating}
  \item{affectionate}{Affectionate round-robin rating}
  \item{knowing}{Do perceivers know the respective targets?}
}
}
\usage{
leipzig
}
\description{
Data from a round-robin study conducted by Richard Rau at the University of Leipzig in 2017.
}
\keyword{datasets}
