% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malzer}
\alias{malzer}
\title{Example dataset for SRM fitting}
\format{
A data frame with 124 rows and 14 variables:
\describe{
  \item{rrgroup}{Id of the round-robin group}
  \item{perceiver}{Id of the perceiver (i.e., actor)}
  \item{Target}{Id of the target (i.e., partner)}
  \item{social}{Social round-robin rating}
}
}
\usage{
malzer
}
\description{
Data from a round-robin study reported in Warner et al. (1979).
}
\keyword{datasets}
