% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm.R
\name{srm}
\alias{srm}
\title{Fit a Social Relations Model to Round-Robin Data}
\usage{
srm(
  formula = NULL,
  g_var = NULL,
  p_var = NULL,
  d_var = NULL,
  data = NULL,
  control = make_args_list(),
  debug = FALSE
)
}
\arguments{
\item{formula}{A formula for the fixed effects}

\item{g_var}{A vector with an identifier for round-robin groups}

\item{p_var}{A vector with identifiers for actor and partner}

\item{d_var}{An optional argument with an identifier for the dyad}

\item{data}{A data frame}

\item{control}{A list of control arguments}

\item{debug}{A control args for debugging, for internal use only}
}
\description{
Fit a Social Relations Model to Round-Robin Data
}
