% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrm-package.R
\docType{package}
\name{tsrm-package}
\alias{tsrm}
\alias{tsrm-package}
\title{tsrm: Fit triadic and dyadic round-robin data}
\description{
Provides functionality for maximum likelihood, 
  restrictced maximum likelihood and Bayesian estimation of 
  Social Relations Model Parameters for Triadic and Dyadic Data.
}
\keyword{internal}
