% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warner}
\alias{warner}
\title{Example dataset for SRM fitting}
\format{
A data frame with 124 rows and 14 variables:
\describe{
  \item{Actor}{Id of the actor (i.e., perceiver)}
  \item{Target}{Id of the partner (i.e., target)}
  \item{Dyad}{Id of the Dyad}
  \item{y1}{First round-robin perception}
  \item{y2}{Second round-robin perception}
  \item{y3}{Third round-robin perception}
}
}
\usage{
warner
}
\description{
Data from a round-robin study reported in Warner et al. (1979).
}
\keyword{datasets}
